[FIX] 플레이어 관련 함수 모듈화
