--- Assets/Scripts/PlayerState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 27be5c20b00b6754ca316145001b9ec0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
